Added the Who'sOnFirst module
Added Who'sOnFirst module with all the sub functions contained by previous modules

The setDIspWord and setFindWord act as setters for setting the instance variables for the display word and the find word to be used by the other methods in the class, like solve.

This module largely uses dicts and lists to map keys to specific values and list of values.

Needs some more error checking as some values might not be set specifically when a function is called